backend/pyproject.toml
[project]
name = "mylists"
version = "2.2.0"
description = ""
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "flask-bcrypt>=1.0.1",
    "flask-caching>=2.3.0",
    "flask-cors>=5.0.0",
    "flask-httpauth>=4.8.0",
    "flask-limiter[redis]>=3.9.2",
    "flask-mail>=0.10.0",
    "flask-marshmallow>=1.2.1",
    "flask-migrate>=4.0.7",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "howlongtobeatpy>=1.0.14",
    "marshmallow>=3.23.2",
    "marshmallow-sqlalchemy>=1.1.0",
    "pillow>=11.0.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "redis>=5.2.1",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "webargs>=8.6.0",
]
